Update Dockerfile and docker-compose networks
The Dockerfile has been updated to include a build argument `BUILDPLATFORM` and to use the `--platform` flag with the base image. This allows for specifying the platform when building the Docker image, which can be useful for cross-platform builds.

In the `docker-compose.yml` file, the `pokemon-basic-network` and `pokeweakness-network` networks have been modified. The `pokemon-basic-network` now explicitly specifies the `bridge` driver and has been given a name. Similarly, the `pokeweakness-network` now also uses the `bridge` driver and has been given a name.